Removes do_cloudflare.
