+++ Tests/Integration/Subscriber/setVarnishPurgeRequestHost.php
@@ -12,22 +12,9 @@ use function WPMedia\Cloudflare\Tests\Integration\getFactory;
  */
 class Test_SetVarnishPurgeRequestHost extends TestCase {
 
-	public function testShouldReturnDefaultWhenCloudflareDisabled() {
-		$options = getFactory()->getContainer( 'options' );
-		$options->set_values( [ 'do_cloudflare' => 0 ] );
-
-		$this->assertSame(
-			'example.org',
-			apply_filters( 'rocket_varnish_purge_request_host', 'example.org' )
-		);
-	}
-
 	public function testShouldReturnDefaultWhenVarnishDisabled() {
 		$options = getFactory()->getContainer( 'options' );
-		$options->set_values( [
-			'do_cloudflare'      => 1,
-			'varnish_auto_purge' => 0,
-		] );
+		$options->set_values( [ 'varnish_auto_purge' => 0 ] );
 
 		$this->assertSame(
 			'example.org',
@@ -37,10 +24,7 @@ class Test_SetVarnishPurgeRequestHost extends TestCase {
 
 	public function testShouldReturnCurrentHostWhenVarnishEnabled() {
 		$options = getFactory()->getContainer( 'options' );
-		$options->set_values( [
-			'do_cloudflare'      => 1,
-			'varnish_auto_purge' => 1,
-		] );
+		$options->set_values( [  'varnish_auto_purge' => 1 ] );
 
 		$this->assertSame(
 			'example.org',
@@ -50,10 +34,7 @@ class Test_SetVarnishPurgeRequestHost extends TestCase {
 
 	public function testShouldReturnCurrentHostWhenFilterTrue() {
 		$options = getFactory()->getContainer( 'options' );
-		$options->set_values( [
-			'do_cloudflare'      => 1,
-			'varnish_auto_purge' => 0,
-		] );
+		$options->set_values( [ 'varnish_auto_purge' => 0 ] );
 
 		add_filter( 'do_rocket_varnish_http_purge', '__return_true' );
