+++ Tests/Unit/Subscriber/autoPurgeByUrl.php
@@ -1,41 +1,22 @@
 <?php
 
-namespace WPMedia\Cloudflare\Tests\Unit\CloudflareSubscriber;
+namespace WPMedia\Cloudflare\Tests\Unit\Subscriber;
 
 use Brain\Monkey\Functions;
 use Mockery;
 use WPMedia\Cloudflare\Cloudflare;
 use WPMedia\Cloudflare\Subscriber;
-use WPMedia\Cloudflare\Tests\Unit\TestCase;
 use WP_Error;
 
 /**
- * @covers WPMedia\Cloudflare\CloudflareSubscriber::auto_purge_by_url
+ * @covers WPMedia\Cloudflare\Subscriber::auto_purge_by_url
  * @group  Cloudflare
  * @group  CloudflareSubscriber
  */
 class Test_AutoPurgeByUrl extends TestCase {
 
-	/**
-	 * Test should not AutoPurge Cloudflare if Cloudflare Addon is disabled.
-	 */
-	public function testShouldNotAutoPurgeByUrlWhenCloudflareIsDisabled() {
-		$mocks = $this->getConstructorMocks( 0, '', '', '' );
-
-		$cloudflare = Mockery::mock( Cloudflare::class );
-		$cloudflare->shouldNotReceive( 'has_page_rule' );
-		$cloudflare->shouldNotReceive( 'purge_by_url' );
-
-		$cloudflare_subscriber = new Subscriber( $cloudflare, $mocks['options_data'], $mocks['options'] );
-
-		$cloudflare_subscriber->auto_purge_by_url( 1, [ '/hello-world' ], '' );
-	}
-
-	/**
-	 * Test should not AutoPurge Cloudflare if Cloudflare Addon is enabled but user has no permissions.
-	 */
-	public function testShouldNotAutoPurgeByUrlWhenCloudflareIsEnabledButNoUserPermissions() {
-		$mocks = $this->getConstructorMocks( 1, '', '', '' );
+	public function testShouldNotAutoPurgeByUrlWhenNoUserPermissions() {
+		$mocks = $this->getConstructorMocks();
 		Functions\when( 'current_user_can' )->justReturn( false );
 
 		$cloudflare = Mockery::mock( Cloudflare::class );
@@ -47,11 +28,8 @@ class Test_AutoPurgeByUrl extends TestCase {
 		$cloudflare_subscriber->auto_purge_by_url( 1, [ '/hello-world' ], '' );
 	}
 
-	/**
-	 * Test should not AutoPurge Cloudflare if Cache Everything Page Rule is not set.
-	 */
 	public function testShouldNotAutoPurgeByUrlWhenNoCacheEverything() {
-		$mocks = $this->getConstructorMocks( 1, '', '', '' );
+		$mocks = $this->getConstructorMocks();
 		Functions\when( 'current_user_can' )->justReturn( true );
 		Functions\when( 'is_wp_error' )->justReturn( true );
 		$wp_error   = Mockery::mock( WP_Error::class );
@@ -64,11 +42,8 @@ class Test_AutoPurgeByUrl extends TestCase {
 		$cloudflare_subscriber->auto_purge_by_url( 1, [ '/hello-world' ], '' );
 	}
 
-	/**
-	 * Test should AutoPurge Cloudflare if Cache Everything Page Rule is set but receives an error.
-	 */
 	public function testShouldAutoPurgeByUrlWhenCacheEverythingButReturnError() {
-		$mocks = $this->getConstructorMocks( 1, '', '', '' );
+		$mocks = $this->getConstructorMocks();
 		Functions\when( 'current_user_can' )->justReturn( true );
 		Functions\when( 'get_current_user_id' )->justReturn( 1 );
 		Functions\expect( 'get_rocket_i18n_home_url' )->once()->with( '' )->andReturn( 'http://example.org/' );
@@ -92,11 +67,8 @@ class Test_AutoPurgeByUrl extends TestCase {
 		$cloudflare_subscriber->auto_purge_by_url( 1, [ '/hello-world' ], '' );
 	}
 
-	/**
-	 * Test should AutoPurge Cloudflare if Cache Everything Page Rule is set.
-	 */
 	public function testShouldAutoPurgeByUrlWhenCacheEverything() {
-		$mocks = $this->getConstructorMocks( 1, '', '', '' );
+		$mocks = $this->getConstructorMocks();
 		Functions\when( 'current_user_can' )->justReturn( true );
 		Functions\when( 'get_current_user_id' )->justReturn( 1 );
 		Functions\when( 'is_wp_error' )->justReturn( false );
@@ -116,54 +88,5 @@ class Test_AutoPurgeByUrl extends TestCase {
 
 		$cloudflare_subscriber = new Subscriber( $cloudflare, $mocks['options_data'], $mocks['options'] );
 		$cloudflare_subscriber->auto_purge_by_url( 1, [ '/hello-world' ], '' );
-	}
-
-	/**
-	 * Get the mocks required by Cloudflare’s constructor.
-	 *
-	 * @since  1.0
-	 * @author Soponar Cristina
-	 * @access private
-	 *
-	 * @param integer $do_cloudflare      - Value to return for $options->get( 'do_cloudflare' ).
-	 * @param string  $cloudflare_email   - Value to return for $options->get( 'cloudflare_email' ).
-	 * @param string  $cloudflare_api_key - Value to return for $options->get( 'cloudflare_api_key' ).
-	 * @param string  $cloudflare_zone_id - Value to return for $options->get( 'cloudflare_zone_id' ).
-	 *
-	 * @return array                      - Array of Mocks
-	 */
-	private function getConstructorMocks( $do_cloudflare = 1, $cloudflare_email = '', $cloudflare_api_key = '', $cloudflare_zone_id = '' ) {
-		$options      = $this->createMock( 'WP_Rocket\Admin\Options' );
-		$options_data = $this->createMock( 'WP_Rocket\Admin\Options_Data' );
-		$map          = [
-			[
-				'do_cloudflare',
-				'',
-				$do_cloudflare,
-			],
-			[
-				'cloudflare_email',
-				null,
-				$cloudflare_email,
-			],
-			[
-				'cloudflare_api_key',
-				null,
-				$cloudflare_api_key,
-			],
-			[
-				'cloudflare_zone_id',
-				null,
-				$cloudflare_zone_id,
-			],
-		];
-		$options_data->method( 'get' )->will( $this->returnValueMap( $map ) );
-
-		$mocks = [
-			'options_data' => $options_data,
-			'options'      => $options,
-		];
-
-		return $mocks;
 	}
 }
