+++ Tests/Unit/TestCase.php
@@ -19,7 +19,6 @@ abstract class TestCase extends WPMediaTestCase {
 		rocket_get_constant( 'WP_ROCKET_VERSION', '3.5' );
 
 		Functions\stubs( [ 'sanitize_text_field' ] );
-
 	}
 
 	protected function getAPIMock() {
